Add grpc client support server streaming

+++ src/Grpc/Server/Response/Stream.php
@@ -70,37 +70,47 @@ class Stream
     /**
      * 写入消息
      * @param mixed $data
-     * @param bool $end
      * @return bool
      */
     public function write(mixed $data): bool
     {
-        if (!$this->withHeader && !$this->withHeader()) {
-            return false;
-        }
-
         if (!$this->response->isWritable()) {
             return false;
         }
 
-        $res = $this->server->send($this->response->fd, $this->frame(
+        $stream = '';
+        // add header
+        if (!$this->withHeader) {
+            $stream .= $this->buildHeader();
+            $this->withHeader = true;
+        }
+        // add message
+        $stream .= $this->frame(
             $data ? Parser::serializeMessage($data) : '',
             self::SW_HTTP2_FRAME_TYPE_DATA,
             self::SW_HTTP2_FLAG_NONE,
             $this->request->streamId
-        ));
+        );
 
-        return $res;
+        return $this->server->send($this->response->fd, $stream);
     }
 
     public function close($status = StatusCode::OK, string $message = ''): bool
     {
-        $res = $this->withHeader(true, [
+        if (!$this->response->isWritable()) {
+            return true;
+        }
+
+        $header = $this->buildHeader(true, [
+            [':status', '200'],
+            ['content-type', 'application/grpc+proto'],
+            ['trailer', 'grpc-status, grpc-message'],
             ['grpc-status', (string)$status],
             ['grpc-message', $message],
         ]);
 
-        $this->response->detach();
+        $res = $this->server->send($this->response->fd, $header);
+        if ($res) $this->response->detach();
 
         return $res;
     }
@@ -111,29 +121,22 @@ class Stream
         return (substr(pack("NccN", strlen($data), $type, $flags, $stream), 1) . $data);
     }
 
-    /**
-     * @param bool $end
-     * @param array $headers
-     * @return bool
-     */
-    public function withHeader(bool $end = false, array $headers = [
+    private function buildHeader(bool $end = false, array $headers = [
         [':status', '200'],
-        ['content-type', 'application/grpc'],
-        ['trailer', 'grpc-status, grpc-message']
-    ]): bool
+        ['content-type', 'application/grpc+proto']
+    ]): string
     {
         try {
             $compressedHeaders = (new HPack())->encode($headers);
         } catch (HPackException) {
-            return false;
+            return '';
         }
-        $res = $this->server->send($this->response->fd, $this->frame(
+
+        return $this->frame(
             $compressedHeaders,
             self::SW_HTTP2_FRAME_TYPE_HEAD,
             $end ? (self::SW_HTTP2_FLAG_END_STREAM | self::SW_HTTP2_FLAG_END_HEADERS) : self::SW_HTTP2_FLAG_END_HEADERS,
             $this->request->streamId
-        ));
-        $this->withHeader = $res;
-        return $res;
+        );
     }
 }
